# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Market API",
        "description": "Market L5 Swagger OpenApi description",
        "contact": {
            "email": "dima8045@ukr.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://market.loc/api",
            "description": "Market API local host"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get list of categories with relation data",
                "operationId": "products",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "The category id specific products category",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Products per page",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Products page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "description": "Products list and pagination parameters",
                                            "properties": {
                                                "current_page": {
                                                    "description": "Current page number",
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "description": "Products array",
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "Product ID",
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "description": "Product name",
                                                                "type": "string"
                                                            },
                                                            "category_id": {
                                                                "description": "Product category ID",
                                                                "type": "integer"
                                                            },
                                                            "image_folder": {
                                                                "description": "Product image folder url",
                                                                "type": "string"
                                                            },
                                                            "unit_id": {
                                                                "description": "Product unit ID",
                                                                "type": "integer"
                                                            },
                                                            "sort_order": {
                                                                "description": "Sort product value",
                                                                "type": "integer"
                                                            },
                                                            "category": {
                                                                "description": "Specified product category",
                                                                "properties": {
                                                                    "id": {
                                                                        "description": "Specified product category id",
                                                                        "type": "integer"
                                                                    },
                                                                    "parent_id": {
                                                                        "description": "Specified parent category id",
                                                                        "type": "integer"
                                                                    },
                                                                    "name": {
                                                                        "description": "Specified product category name",
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "product_images": {
                                                                "description": "Specified product images",
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "description": "Specified product category id",
                                                                            "type": "integer"
                                                                        },
                                                                        "parent_id": {
                                                                            "description": "Specified parent category id",
                                                                            "type": "integer"
                                                                        },
                                                                        "name": {
                                                                            "description": "Specified product category name",
                                                                            "type": "string"
                                                                        },
                                                                        "alt": {
                                                                            "description": "Specified product alternative name",
                                                                            "type": "string"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "unit": {
                                                                "description": "Product unit",
                                                                "properties": {
                                                                    "id": {
                                                                        "description": "Prodict unit ID",
                                                                        "type": "integer"
                                                                    },
                                                                    "name": {
                                                                        "description": "Prodict unit name",
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "last_page": {
                                                    "description": "Last page number",
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "description": "Url to current path",
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "description": "Products per page",
                                                    "type": "string"
                                                },
                                                "total": {
                                                    "description": "Total products number",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}